requirements.txt
Django==1.11.16
djangorestframework==3.8.2
mongoengine==0.15.3
pymongo==3.7.1
pytz==2018.5
redis==2.10.6
six==1.11.0